Refactor memcr-client code to create libmemcrclient.so licensed as LGPL2.1
