+++ COPYING
@@ -31,3 +31,28 @@ General Public License for more details.
 You should have received a copy of the GNU General Public License
 along with this library; if not, write to the Free Software Foundation,
 Inc., 51 Franklin St, Fifth Floor, Boston, MA 02110-1301, USA.
+
+================================================================================
+
+memcrclient_proto.h and libmemcrclient shared libarary are licensed by
+Liberty Global Service B.V. under the GNU Lesser General Public License
+version 2.1 only, as per COPYING.LGPL2.1.
+
+Please use the following copyright for source for libmemcrlient shared library
+source code:
+
+Copyright (C) <YEAR>  <NAME>
+
+This library is free software; you can redistribute it and/or
+modify it under the terms of the GNU Lesser General Public
+License as published by the Free Software Foundation; either
+version 2.1 of the License, or (at your option) any later version.
+
+This library is distributed in the hope that it will be useful,
+but WITHOUT ANY WARRANTY; without even the implied warranty of
+MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
+Lesser General Public License for more details.
+
+You should have received a copy of the GNU Lesser General Public
+License along with this library; if not, see <https://www.gnu.org/licenses/>.
+
